Squiz/FunctionSpacing: iterate on previous fix for attributes
Turns out the "silence the "before" error if the thing before was another function" functionality still wasn't working correctly.

And if there is a blank line between the first thing before the first "preamble" for the function and the declaration itself, the sniff would also disregard the "preamble" and insert the blank lines straight above the function declaration.

Fixed now.

Includes tests proving the bug.

Includes one minor change in the existing tests, where the recognition of "the thing before is also a function" no longer works.
This is a specific edge case when there are two functions declared on the same line and, in my opinion, this tiny regression is a valid trade-off for correctly handling functions with attributes, which is much more commonly found in code.