PHP 8.4 | File::getMemberProperties(): allow for final properties (#834)
Update File::getMemberProperties() to report when a property is marked as final. Update existing tests (which are all for non-final properties) to expect that the property not be marked as final, and add a new test case for a final that should be, and is, marked as final.

Part of #734