Squiz/ClassFileName: expand the tests
This commit adds some extra tests for the sniff.
* Document handling of unfinished OO declarations (which contain enough info to still action).
* Safeguard that comments in unexpected places doesn't cause problems.

As this sniff looks at the file name of the file containing the tests, we cannot use the _normal_ `SniffNameUnitTest.#.inc` naming conventions for extra tests files, so the `getTestFiles()` method from the `AbstractSniffUnitTest` class needs to be overloaded to retrieve the test case files in a slightly different way.