Correct file/class name relation in sniff doc
Per PR #845, the filename must be modified to match the class name,
since class names are much more restrictive than filenames and the
inverse is not always possible.

This clarifies in the example titles that the fix for the invalid code
is to modify the filename rather than the class name.