[Docs] Add XML doc for Squiz.Classes.ClassFileName sniff (#843)
* Add XML doc for Squiz\Classes\ClassFileName sniff

* [Doc] Emphasize sniff focus for ClassFileName

Adjusts the `<em>` tags in the Squiz.Classes.ClassFileName sniff XML
docs to better highlight the parts of the code samples that illustrate
the purpose of the sniff.

* Correct file/class name relation in sniff doc

Per PR #845, the filename must be modified to match the class name,
since class names are much more restrictive than filenames and the
inverse is not always possible.

This clarifies in the example titles that the fix for the invalid code
is to modify the filename rather than the class name.