Ruleset: bug fix - correctly handle empty array property setting
While already handled correctly for inline properties set via `phpcs:set`, setting a property to an empty array from a ruleset file was not handled correctly until now.

As things were, when setting an array property from the ruleset to an empty array like so:
```xml
    <rule ref="Standard.Category.SniffName">
        <properties>
            <property name="arrayProperty" type="array"/>
        </properties>
    </rule>
```

 the property value would be set to:
```php
public $arrayProperty = [
    0 => '',
];
```
... while the expected behaviour would be:
```php
public $arrayProperty = [];
```

While it is not expected that this type of property setting is encountered a lot in the wild, it should still work correctly.

Fixed now.

Includes test safeguarding the fix.