File::getMemberProperties(): removed parse error warning
The `File::getMemberProperties()` method used to be inconsistent in how it handled variable tokens which were not property declarations.
* In most cases ("normal" variable, function parameter), it would throw a `RuntimeException` _"'$stackPtr is not a class member var'"_.
* However, for "properties" declared in an `enum` or `interface` construct, it would register a warning about a possible parse error and return an empty array.

This parse error warning has now been removed.

As, as of PHP 8.4, declaring (hooked) properties in an interface is no longer a parse error (see the [Property Hooks RFC](https://wiki.php.net/rfc/property-hooks)), properties declared in interfaces will now be analyzed by the function, like any other property, and will return an array of information about the property.

For "properties" declared in enums, which is still not allowed in PHP, the method will throw the `RuntimeException` _"'$stackPtr is not a class member var'"_.

Includes updated unit tests for the `File::getMemberProperties()` method.

Includes a review of all uses of the `File::getMemberProperties()` method in PHPCS native sniffs and updating the code where necessary.
Includes adding/updating tests with properties in interfaces and enums for each of those sniffs.

Closes squizlabs/PHP_CodeSniffer 2455

Also related to 734

Co-authored-by: Greg Sherwood <gsherwood@squiz.net>