Config: drop support for setting language via `--extensions` CLI arg
This has been implemented slightly differently from the original PHPCS 4.0 branch implementation.
1. The array format of the `Config::$extensions` property has not been changed.
    The format is an associative array with the extensions as keys and the tokenizer type as the value.
    Originally, the format was changed to a list with the extensions as values.
    While information on the tokenizer type now no longer has any value as the only supported tokenizer is PHP, changing the array format could break integrations in unexpected ways, so I have decided against this.
2. I have also elected to throw an error when the passed extensions contain the `/tokenizer` switch, but only if the tokenizer is set to something other than PHP.
    In the original implementation, no information was provided to the user and all extensions values would be accepted and expected to be acceptable, including values which contained the `/tokenizer` switch.

Note: users can still pass extensions for unsupported file types. This is no different than before, either in the previous implementation or in PHPCS 3.x. I mean, users could already pass `html` as an extension and it would be scanned as if it were PHP. Whether the results of such scans would be useful is doubtful, but that's the user's responsibility.
So if the user passes `--extensions=php,js`, as of PHPCS 4.0, javascript files will be scanned as if they were PHP.

Ref:
* squizlabs/PHP_CodeSniffer 2448

Co-authored-by: Greg Sherwood <gsherwood@squiz.net>