Tests: remove work-arounds for setup/teardown methods
As of PHPUnit 8, the `setUpBeforeClass()`, `setUp()`, `tearDown()` and `tearDownAfterClass()` method were required to have a `void` return type declaration.

This could previously not be applied as the test suite had to stay compatible with PHP < 7.2, so as a work-around, the `@beforeClass`, `@before`, `@after` and `@afterClass` annotations were used for these setup/teardown methods.

As support for PHPUnit < 8 has now been dropped, this work-around is no longer needed and the methods can use the more customary PHPUnit fixture method signature again.

Co-authored-by: Greg Sherwood <gsherwood@squiz.net>