GH Actions/coverage: join some steps
The "coverage" job basically runs the tests in three different ways:
1. Normal test run on Linux.
2. CBF specific tests in CBF mode on Linux.
3. Windows specific tests on Windows.

For these three test runs, the script contained 6 steps as whether or not the coverage cache should be used depends on the PHPUnit version.

This commit condenses the 6 steps to 3 steps by using a command inline expression instead of step conditions.