Tests: remove work-arounds for `--sniffs` not being respected
Now that the `--sniffs` CLI argument will be respected in the tests, a number of work-arounds previously introduced can be removed.

Note: there may be a few more work-arounds which can be removed, like removing some fixture rulesets in favour of test inline sniff selection, but this should be a good first step.