GH Actions: auto-remove label tweaks
PR 897 made a change to the "auto remove labels" workflow, but that change couldn't be tested until merged as it used the `pull_request_target` event, which always reads the workflows from the `master` branch and not from a PR branch.

I've now tested the change and while the _new_ addition looks to be correct, one of the pre-existing jobs _also_ ran, while it shouldn't have run.

This new commit attempts to fix that.

Again, we can't test this until merged, but will test once it has been merged.