PHP 8.1 | Tokenizer/PHP: fix nullable type tokenization for readonly properties
The `?` in a type for a `readonly` property declared without visibility, would be tokenized as `T_INLINE_THEN`, not `T_NULLABLE`.

Fixed now.

Includes perfunctory test.
Realistically, a lot more tests are needed for the `T_NULLABLE` vs `T_INLINE_THEN` tokenization, but at least, this is a start and covers the current change.