Various sniffs + File methods: remove redundant code/simplify for T_USE parenthesis owner
As `T_USE` tokens are now parenthesis owners and will have the `parenthesis_*` keys, sniffs do not need to token walk anymore to find the parentheses (or the `T_USE` token).

It has been confirmed that the changes in this commit are all covered by pre-existing tests.

Replaces squizlabs/PHP_CodeSniffer 3338
Closes 43