Tokenizer::createTokenMap(): stabilize code for parenthesis owners
_I didn't dare to make this change while JS/CSS were still supported, but this should be a safe change now._

For all PHP parenthesis owners, the open parenthesis MUST be the next non-empty token.
There is only one exception to this rule and that is the `T_FUNCTION` keyword where a `T_STRING` function name and potentially a `&` for return by reference are also allowed between.

If the parenthesis owner is not the next valid token, the token should not be marked as a parenthesis owner, not get the other parenthesis keys.

Now, while for valid code the above should always be true. For invalid code and a few edge cases, such as anonymous classes without parentheses, it is currently possible for the tokenizer to walk too far and identify a completely unrelated arbitrary open parenthesis as the parenthesis belonging to a parenthesis owner.

The change in this commit should prevent that.