:sparkles: New Generic.WhiteSpace.GotoTargetSpacing sniff
What with the colon for a goto target label no longer being tokenized to be included in the label token, let's allow for people to enforce no space between the label and the colon.

This commit introduces a new sniff to do just that.

Includes tests.
Includes documentation.

Related to 185