Tokenizer/PHP: `T_GOTO_LABEL` no longer contains colon
When support for the `goto` construct was introduced, it was elected to tokenize the "goto target label" as a separate `T_GOTO_LABEL` token, including the colon following it.

This is wrong as PHP does not enforce for the colon to be the next token and allows whitespace and comments between the goto label and the colon. In which case, the tokenization in PHPCS would be incorrect as the label would stay a `T_STRING` instead of becoming `T_GOTO_LABEL`.

This commit fixes this.

It also introduces a new `T_GOTO_COLON` token and re-tokenizes the colon (separate from the label) to `T_GOTO_COLON`.
This choice was made so as not to confuse the Tokenizer (nor sniffs) with yet another usage for `T_COLON`, which is an ambiguous enough token as it is (alternative control structures, switch - case statements, return type statements, enum types, ternaries etc).

This change is covered by the existing (updated) tests.
Includes one minor update to a sniff to allow for this change.

Refs:
* https://www.php.net/manual/en/control-structures.goto.php
* https://3v4l.org/ccbVD

Fixes squizlabs/PHP_CodeSniffer 3161
Fixes 185