Generic/FunctionCallArgumentSpacing: update XML doc to remove info about spacing around equal sign
The FunctionCallArgumentSpacing sniff used to check for spacing around
the equal sign in the arguments of a function declaration. But this was
removed in https://github.com/squizlabs/PHP_CodeSniffer/pull/2399
(commit feeda473f045825c1caa213c3b7ab6dc6a99b746).

This commit updates the sniff XML doc to reflect this change.