Generic/FunctionCallArgumentSpacing: fix code examples in the XML doc
The code examples in the XML documentation of this sniff were incorrect
as the sniff does not check spacing in function argument definitions. It
only checks spacing in function and method calls.

As far as I could check this has always been the case, or at least it
has been the case for a long time. Since tests were added for this
sniff when it was part of the PEAR standard and called
MethodCallArgumentSpacing, it ignored function definitions and checked
only function calls:

https://github.com/squizlabs/PHP_CodeSniffer/commit/7b84951474ff24d394eb71e42ad79db99cf776bb#diff-56e2f69852c3071fbf86d20fbb49f38dae17cff8731154d8b1b7495c4ad8eeb0R24-R27

Here is the sniff code where it ignores function definitions:

https://github.com/PHPCSStandards/PHP_CodeSniffer/blob/57a9ed734e2ab06c062fe0e463745eab483cf722/src/Standards/Generic/Sniffs/Functions/FunctionCallArgumentSpacingSniff.php#L55-L66