Tokenizer/PHP: fix context sensitive keywords as goto labels
Previously quite some work has already been done to ensure that keywords which can be used in a non-keyword context would be tokenized as `T_STRING` in those situations.

The `goto` context was, however, not taken into account.

It should be noted that none of the official PHP keywords are allowed to be used as `goto` labels, so all cases covered by the `ContextSensitiveKeywordsGotoTest` are in actual fact parse errors.

Having said that, those keywords being tokenized as the keyword, instead of as `T_STRING`, could seriously confuse sniffs, which is why, IMO, they should be (re-)tokenized to `T_STRING`.

For the record: the "other" context sensitive keywords - `parent`, `self`, `true`, `false` and `null` are actually allowed as labels for `goto` statements, even though this makes for some pretty confusing code. Example: https://3v4l.org/dcbmm

Either way, fixed now (for the most part).

Note: This fix does not include fixing the (re-)tokenization of keywords to `T_GOTO_LABEL` for the "goto target statement". I'm not sure this can be done reliably without breaking other things, so that's considered outside the scope of this PR.

Includes plenty of tests to safeguard the change.

Related to 185