Generic/OpeningFunctionBraceKernighanRitchie: improve XML doc
- `<documentation>` title matches the sniff name.
- Don't mention the Kernighan/Ritchie style in the `<standard>`
description as according to Wikipedia this style uses the opening brace
for functions on a separate line and not on the same line (https://en
.wikipedia.org/wiki/Indentation_style#K&R).
- I believe it is more precise to say that the sniff checks if the
opening brace is on the same line as the closing parenthesis and not the
function declaration as the function declaration can be multi-line.
- Mention in the `<standard>` description that there must be no content
on the line after the opening brace.
- Improve the title of the `<code>` blocks.