ProcessRulesetBrokenRulesetTest: more tweaks for different libxml output
Looks like there are some more libxml quirks to take into account with different message content on different OSes.

This updates the tests to allow for them to pass on PHP < 8.0 with libxml 2.9.x and lower on Windows.

Also adds the `@group Windows` annotation to ensure these tests are also run on Windows in CI.