PHP 7.4 | Tokenizer/PHP: handle PHP tag at end of file consistently
Prior to PHP 7.4, a PHP open tag at the end of a file was not tokenized correctly in PHP itself.

From the PHP 7.4 changelog:
> `<?php` at the end of the file (without trailing newline) will now be
> interpreted as an opening PHP tag. Previously it was interpreted either as
> `<? php` and resulted in a syntax error (with short_open_tag=1) or was
> interpreted as a literal `<?php` string (with short_open_tag=0).

This commit makes the tokenization of PHP open tags at the end of a file consistent in all PHP versions.

Includes tests.
Includes adding a few extra test builds which test with `short_open_tag=On` for various PHP versions.

Refs:
* https://www.php.net/manual/en/migration74.incompatible.php#migration74.incompatible.core.php-tag
* https://github.com/php/php-src/blob/30de357fa14480468132bbc22a272aeb91789ba8/UPGRADING#L37-L40