Squiz/OperatorBracket: stability tweak
The `$allowed` tokens array is joined after declaration with the `Tokens::$operators` array. This is only reliable if the keys in both arrays are unique (or intended to be overwritten).
However, with the `$allowed` array being numerically indexed and PHP native tokens also being integers, this was not the case and could lead to one of the tokens in the original array (unintentionally) being disregarded/overwritten by one of the tokens in the `Tokens::$operators` array.

Fixed now, by making setting the tokens constants both as the key as well as the value, making the entries in the array suitable for an array join operation.