PSR2/NamespaceDeclaration: remove redundant condition
What with the change in the namespaced names tokenization, a sniff listening for `T_NAMESPACE` will no longer receive tokens for namespace relative names as those no longer use that token.

Keeping the existing tests for this in place as those are still a good safeguard to prevent a regression slipping in.